fix(DigitalPin): fix data types

+++ lib/Commons/Commons.h
@@ -30,7 +30,7 @@ namespace Victoria {
 
   struct AppModel {
     String name;
-    int ledPin = 2;
+    uint8_t ledPin = 2;
     uint8_t ledOnValue = 0; // LOW
   };
